+++ ingeniamotion/fsoe.py
@@ -8,6 +8,7 @@ from fsoe_master.fsoe_master import (
     DictionaryItem,
     DictionaryItemInputOutput,
     MasterHandler,
+    State,
     StateData,
 )
 from ingenialink.enums.register import REG_ACCESS, REG_DTYPE
@@ -138,7 +139,7 @@ class FSoEMasterHandler:
         state_reached = False
         init_time = time.time()
         while not state_reached:
-            state_reached = self.__master_handler.state == StateData
+            state_reached = self.state == StateData
             if timeout and (init_time + timeout) < time.time():
                 raise IMTimeoutError("The FSoE Master did not reach the Data state")
 
@@ -162,6 +163,11 @@ class FSoEMasterHandler:
     def safety_slave_pdu_map(self) -> TPDOMap:
         """The PDOMap used for the Safety Slave PDU."""
         return self.__safety_slave_pdu
+
+    @property
+    def state(self) -> State:
+        """Get the FSoE master state."""
+        return self.__master_handler.state
 
 
 class PDUMapper:
@@ -358,6 +364,19 @@ class FSoEMaster:
         """
         master_handler = self.__handlers[servo]
         master_handler.wait_for_data_state(timeout)
+
+    def get_fsoe_master_state(self, servo: str = DEFAULT_SERVO) -> State:
+        """Get the servo's FSoE master handler state.
+
+        Args:
+            servo: servo alias to reference it. ``default`` by default.
+
+        Returns:
+            The servo's FSoE master handler state.
+
+        """
+        master_handler = self.__handlers[servo]
+        return master_handler.state
 
     def _delete_master_handler(self, servo: str = DEFAULT_SERVO) -> None:
         """Delete the master handler instance